# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基於Gin + Vue + Element UI前後端分離的工單系統的接口文檔\n添加qq群: 1127401830 進入技術交流群 請備注，謝謝！",
        "title": "ferry API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/lanyulei/ferry/blob/master/LICENSE"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "獲取token",
                "consumes": [
                    "application/json"
                ],
                "summary": "登陸",
                "parameters": [
                    {
                        "description": "Add account",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"expire\": \"2019-08-07T12:45:48+08:00\", \"token\": \".eyJleHAiOjE1NjUxNTMxNDgsImlkIjoiYWRtaW4iLCJvcmlnX2lhdCI6MTU2NTE0OTU0OH0.-zvzHvbg0A\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "獲取token",
                "consumes": [
                    "application/json"
                ],
                "summary": "退出登錄",
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"msg\": \"成功退出系統\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/sd/disk": {
            "get": {
                "description": "服務器硬盤使用量 DiskCheck checks the disk usage.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "summary": "服務器硬盤使用量",
                "responses": {
                    "200": {
                        "description": "OK - Free space: 16321MB (15GB) / 51200MB (50GB) | Used: 31%",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "429": {
                        "description": "WARNING",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "CRITICAL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Login": {
            "type": "object",
            "required": [
                "code",
                "password",
                "username",
                "uuid"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}